Support extensions in openapiv2_operation options
Extensions as specified in OpenAPI are not supported by grpc-gateway.
This adds an extensions key in the openapiv2_operation option for
protoc-swagger-gen as a string-keyed map with Any values. When
marshaling the Operation we merge the serialization of the extensions
map and the Operation struct to form one JSON object in the generated
OpenAPI document.